+++ changelog.md
@@ -1,3 +1,6 @@
+## Upcoming
+* Disable stone tools feature now also disables stone shields from Shields+
+
 ## 5.4.1.2
 * Sweet Berries Patches are now 3x rarer
 * Fixed Nether Fog changes not working